Rename dir/module to make more sense
