+++ app/leagueConfiguration/AmazingRace_35.tsx
